Moved scorekeeping to GameLogic where it is being used anyway.
removed Menuselect for readability, its back in Program

Removed returning a null because its the devil, replaced it with "" because i dont need it to return anything.

This directly affects ToString and is a refactoring that generally makes sense, Tostring now calls a Method to remove the asterisks and print em, which means it supports as many answers as you want.

Restructured some things in UI to fit the new order.